feat: refactor: 패키지명 변경과 필요없는 파일 삭제
