+++ backoffice-api/src/main/java/com/fastcampus/backoffice/BackofficeApplication.java
@@ -3,7 +3,7 @@ package com.fastcampus.backoffice;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-@SpringBootApplication
+@SpringBootApplication(scanBasePackages = "com.fastcampus.backoffice")
 public class BackofficeApplication {
     public static void main(String[] args) {
         SpringApplication.run(BackofficeApplication.class, args);
