--- backoffice-api/src/main/java/com/fastcampus/backofficeapi/BackofficeApiApplication.java
@@ -1,13 +1,0 @@
-package com.fastcampus.backofficeapi;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class BackofficeApiApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(BackofficeApiApplication.class, args);
-    }
-
-}
